Move MessageWithBytes into separate file
